Updates to build on Windows 10/.NET 4.0
You need to copy the BAPS.dll and all other files in that directory to the BAPSServerService build directory for it to launch.
Setup still doesn't build.
Required the Windows SDK v10.0.15.063.0, MFC and ATL support, and VC++ 2015.3 v140 toolset from the Visual Studio installer (VS 2017).

Added checks to ConfigManager.cpp to allow BAPS to be run without a USB controller, even if the controller configuration option is set as a debug workaround.